+++ demo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "predictive-textarea-demo",
+  "name": "demo",
   "private": true,
   "type": "module",
   "scripts": {
