+++ dev/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "predictive-textarea-dev",
+  "name": "dev",
   "type": "module",
   "private": true,
   "scripts": {
@@ -18,7 +18,7 @@
     "express": "^5.1.0",
     "lucide-react": "^0.344.0",
     "openai": "^4.95.1",
-    "predictive-textarea": "file:../",
+    "predictive-textarea": "file:../predictive-textarea",
     "prism-react-renderer": "^2.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
