+++ package.json
@@ -1,63 +1,16 @@
 {
-  "name": "predictive-textarea",
-  "version": "0.1.1",
-  "description": "A rich text input with AI-powered autocomplete content predictions for React",
-  "type": "module",
-  "main": "./dist/index.umd.js",
-  "module": "./dist/index.es.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
+  "name": "predictive-textarea-monorepo",
+  "private": true,
+  "workspaces": [
+    "predictive-textarea",
+    "dev",
+    "demo"
   ],
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.es.js",
-      "require": "./dist/index.umd.js",
-      "default": "./dist/index.es.js"
-    }
-  },
   "scripts": {
-    "build": "tsup src/index.ts --format cjs,esm --dts",
-    "dev": "cd dev && npm run dev",
-    "lint": "eslint src/**/*.ts*",
-    "clean": "rm -rf dist",
-    "demo": "cd demo && npm run dev",
-    "dev:api": "cd dev && npm run dev:api",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
-    "prepare-version": "tsx scripts/release-branch.ts prepare-version",
-    "create-release": "tsx scripts/release-branch.ts create-release"
+    "build": "yarn workspace predictive-textarea build",
+    "dev": "yarn workspace predictive-textarea-dev dev",
+    "demo": "yarn workspace demo dev",
+    "watch": "yarn workspace predictive-textarea watch"
   },
-  "peerDependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@types/jest": "^29.0.0",
-    "@types/node": "^20.0.0",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@vitejs/plugin-react": "^4.4.0",
-    "dotenv": "^16.5.0",
-    "execa": "^9.5.2",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "octokit": "^4.1.3",
-    "ora": "^8.2.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "ts-jest": "^29.1.0",
-    "ts-node": "^10.9.2",
-    "tsup": "^8.4.0",
-    "typescript": "^5.0.0",
-    "vite": "^5.4.18",
-    "vite-plugin-dts": "^3.9.1"
-  },
-  "dependencies": {
-    "trie-search": "^2.2.0"
-  }
+  "packageManager": "yarn@4.9.1+sha512.f95ce356460e05be48d66401c1ae64ef84d163dd689964962c6888a9810865e39097a5e9de748876c2e0bf89b232d583c33982773e9903ae7a76257270986538"
 }
