+++ predictive-textarea/package.json
@@ -1,0 +1,64 @@
+{
+  "name": "predictive-textarea",
+  "version": "0.1.1",
+  "description": "A rich text input with AI-powered autocomplete content predictions for React",
+  "type": "module",
+  "main": "./dist/index.umd.js",
+  "module": "./dist/index.es.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.es.js",
+      "require": "./dist/index.umd.js",
+      "default": "./dist/index.es.js"
+    }
+  },
+  "scripts": {
+    "build": "tsup src/index.ts --format cjs,esm --dts",
+    "watch": "tsup src/index.ts --format cjs,esm --dts --watch",
+    "dev": "cd dev && npm run dev",
+    "lint": "eslint src/**/*.ts*",
+    "clean": "rm -rf dist",
+    "demo": "cd demo && npm run dev",
+    "dev:api": "cd dev && npm run dev:api",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "prepare-version": "tsx scripts/release-branch.ts prepare-version",
+    "create-release": "tsx scripts/release-branch.ts create-release"
+  },
+  "peerDependencies": {
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@types/jest": "^29.0.0",
+    "@types/node": "^20.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@vitejs/plugin-react": "^4.4.0",
+    "dotenv": "^16.5.0",
+    "execa": "^9.5.2",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "octokit": "^4.1.3",
+    "ora": "^8.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "ts-jest": "^29.1.0",
+    "ts-node": "^10.9.2",
+    "tsup": "^8.4.0",
+    "typescript": "^5.0.0",
+    "vite": "^5.4.18",
+    "vite-plugin-dts": "^3.9.1"
+  },
+  "dependencies": {
+    "trie-search": "^2.2.0"
+  }
+}
