+++ predictive-textarea/src/components/content-prediction.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { cn } from '@/utils/cn'
+import { cn } from '../utils/cn'
 
 
 interface ContentPredictionProps {
