+++ predictive-textarea/src/components/predictive-textarea.tsx
@@ -1,9 +1,9 @@
 import React, { useRef, SyntheticEvent, KeyboardEvent, FocusEvent, useState, useEffect, CSSProperties } from 'react'
 
-import { useCaretPosition } from '@/hooks/use-caret-position'
-import { useContentPrediction } from '@/hooks/use-content-prediction'
-import { PredictiveTextareaProps } from '@/types'
-import { cn } from '@/utils/cn'
+import { useCaretPosition } from '../hooks/use-caret-position'
+import { useContentPrediction } from '../hooks/use-content-prediction'
+import { PredictiveTextareaProps } from '../types'
+import { cn } from '../utils/cn'
 
 import { ContentPrediction } from './content-prediction'
