+++ predictive-textarea/src/hooks/use-caret-position.ts
@@ -1,6 +1,6 @@
 import { RefObject, useCallback } from 'react'
 
-import { isCaretAtLineEnd } from '@/utils/is-caret-at-line-end'
+import { isCaretAtLineEnd } from '../utils/is-caret-at-line-end'
 
 import { getTextUpUntilCaretInElement } from '../utils/get-text-up-until-caret-in-element'
