+++ predictive-textarea/src/hooks/use-content-prediction.ts
