+++ predictive-textarea/src/index.ts
