+++ predictive-textarea/src/utils/get-text-up-until-caret-in-element.ts
