+++ predictive-textarea/src/utils/hash.ts
