+++ predictive-textarea/src/utils/is-caret-at-line-end.ts
