+++ predictive-textarea/tests/hooks/use-content-prediction.test.tsx
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react'
 
-import { useContentPrediction } from '../../src/hooks/use-content-prediction'
+import { useContentPrediction } from '../../package/src/hooks/use-content-prediction'
 
 // Create a mock for the trie-search library
 jest.mock('trie-search', () => {
