+++ predictive-textarea/tests/utils.test.ts
@@ -1,6 +1,6 @@
-import { isCaretAtLineEnd } from '@/utils/is-caret-at-line-end'
+import { isCaretAtLineEnd } from '../package/src/utils/is-caret-at-line-end'
 
-import { getTextUpUntilCaretInElement } from '../src/utils/get-text-up-until-caret-in-element'
+import { getTextUpUntilCaretInElement } from '../package/src/utils/get-text-up-until-caret-in-element'
 
 
 describe('Autocomplete Utils', () => {
