+++ predictive-textarea/tests/utils/get-text-up-until-caret-in-element.test.ts
@@ -1,4 +1,4 @@
-import { getTextUpUntilCaretInElement } from '@/utils/get-text-up-until-caret-in-element'
+import { getTextUpUntilCaretInElement } from '../../package/src/utils/get-text-up-until-caret-in-element'
 
 
 describe('getTextUptilCaretInElement', () => {
