+++ predictive-textarea/tsconfig.json
@@ -28,9 +28,12 @@
       }
     ],
     "paths": {
-      "@/*": [
-        "./src/*"
+      "predictive-textarea": [
+        "../predictive-textarea/src"
       ],
+      "predictive-textarea/*": [
+        "../predictive-textarea/src/*"
+      ]
     },
     "target": "ES2022",
     "noUncheckedIndexedAccess": true
